[General] Change underlying type of the WidgetType to uint8_t

+++ components/openmw-mp/Base/BasePlayer.hpp
@@ -190,7 +190,7 @@ namespace mwmp
         {
             int32_t id;
             short width, height;
-            enum class WidgetType: int
+            enum class WidgetType: uint8_t
             {
                 Button,
                 Editbox,
